requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
line-bot-sdk==3.17.1
gunicorn==20.1.0
google-generativeai